+++ k8sutil/pkg/controller/obj/api/networkingv1/networkpolicy.go
@@ -1,12 +1,9 @@
 package networkingv1
 
 import (
-	"context"
-
 	"github.com/puppetlabs/leg/k8sutil/pkg/controller/obj/helper"
 	"github.com/puppetlabs/leg/k8sutil/pkg/controller/obj/lifecycle"
 	networkingv1 "k8s.io/api/networking/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -15,46 +12,20 @@ var (
 )
 
 type NetworkPolicy struct {
+	*helper.NamespaceScopedAPIObject
+
 	Key    client.ObjectKey
 	Object *networkingv1.NetworkPolicy
 }
 
-var _ lifecycle.Deleter = &NetworkPolicy{}
-var _ lifecycle.LabelAnnotatableFrom = &NetworkPolicy{}
-var _ lifecycle.Loader = &NetworkPolicy{}
-var _ lifecycle.Ownable = &NetworkPolicy{}
-var _ lifecycle.Persister = &NetworkPolicy{}
-
-func (np *NetworkPolicy) Delete(ctx context.Context, cl client.Client, opts ...lifecycle.DeleteOption) (bool, error) {
-	return helper.DeleteIgnoreNotFound(ctx, cl, np.Object, opts...)
-}
-
-func (np *NetworkPolicy) LabelAnnotateFrom(ctx context.Context, from metav1.Object) {
-	helper.CopyLabelsAndAnnotations(&np.Object.ObjectMeta, from)
-}
-
-func (np *NetworkPolicy) Load(ctx context.Context, cl client.Client) (bool, error) {
-	return helper.GetIgnoreNotFound(ctx, cl, np.Key, np.Object)
-}
-
-func (np *NetworkPolicy) Owned(ctx context.Context, owner lifecycle.TypedObject) error {
-	return helper.Own(np.Object, owner)
-}
-
-func (np *NetworkPolicy) Persist(ctx context.Context, cl client.Client) error {
-	if err := helper.CreateOrUpdate(ctx, cl, np.Object, helper.WithObjectKey(np.Key)); err != nil {
-		return err
-	}
-
-	np.Key = client.ObjectKeyFromObject(np.Object)
-	return nil
+func makeNetworkPolicy(key client.ObjectKey, obj *networkingv1.NetworkPolicy) *NetworkPolicy {
+	np := &NetworkPolicy{Key: key, Object: obj}
+	np.NamespaceScopedAPIObject = helper.ForNamespaceScopedAPIObject(&np.Key, lifecycle.TypedObject{GVK: NetworkPolicyKind, Object: np.Object})
+	return np
 }
 
 func (np *NetworkPolicy) Copy() *NetworkPolicy {
-	return &NetworkPolicy{
-		Key:    np.Key,
-		Object: np.Object.DeepCopy(),
-	}
+	return makeNetworkPolicy(np.Key, np.Object.DeepCopy())
 }
 
 func (np *NetworkPolicy) AllowAll() {
@@ -78,17 +49,11 @@ func (np *NetworkPolicy) DenyAll() {
 }
 
 func NewNetworkPolicy(key client.ObjectKey) *NetworkPolicy {
-	return &NetworkPolicy{
-		Key:    key,
-		Object: &networkingv1.NetworkPolicy{},
-	}
+	return makeNetworkPolicy(key, &networkingv1.NetworkPolicy{})
 }
 
 func NewNetworkPolicyFromObject(obj *networkingv1.NetworkPolicy) *NetworkPolicy {
-	return &NetworkPolicy{
-		Key:    client.ObjectKeyFromObject(obj),
-		Object: obj,
-	}
+	return makeNetworkPolicy(client.ObjectKeyFromObject(obj), obj)
 }
 
 func NewNetworkPolicyPatcher(upd, orig *NetworkPolicy) lifecycle.Persister {